+++ lib/src/style/react_circle_tab_style.dart
@@ -2,8 +2,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import '../../convex_bottom_bar.dart';
-import 'transition_container.dart';
 import 'blend_image_icon.dart';
+import 'transition_container.dart';
 
 /// Convex shape is moved after selection
 class ReactCircleTabStyle extends DelegateBuilder {
@@ -12,6 +12,7 @@ class ReactCircleTabStyle extends DelegateBuilder {
   final Color color;
   final Color backgroundColor;
   final Curve curve;
+  final margin = (ACTION_LAYOUT_SIZE - ACTION_INNER_BUTTON_SIZE) / 2;
 
   ReactCircleTabStyle({
     this.items,
@@ -26,32 +27,25 @@ class ReactCircleTabStyle extends DelegateBuilder {
     var item = items[index];
     if (active) {
       final item = items[index];
-      return Stack(
-        alignment: Alignment.center,
-        children: <Widget>[
-          TransitionContainer.scale(
-            child: Container(
-              width: ACTION_LAYOUT_SIZE,
-              height: ACTION_LAYOUT_SIZE,
-              padding: EdgeInsets.all(
-                  (ACTION_LAYOUT_SIZE - ACTION_INNER_BUTTON_SIZE) / 2),
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: active ? activeColor : color,
-              ),
-              child: BlendImageIcon(
-                active ? item.activeIcon ?? item.icon : item.icon,
-                size: ACTION_INNER_BUTTON_SIZE,
-                color: item.blend ? backgroundColor : null,
-              ),
-            ),
-            curve: curve,
+      return TransitionContainer.scale(
+        child: Container(
+          width: ACTION_LAYOUT_SIZE,
+          height: ACTION_LAYOUT_SIZE,
+          margin: EdgeInsets.all(margin),
+          decoration: BoxDecoration(
+            shape: BoxShape.circle,
+            color: active ? activeColor : color,
           ),
-        ],
+          child: BlendImageIcon(
+            active ? item.activeIcon ?? item.icon : item.icon,
+            size: ACTION_INNER_BUTTON_SIZE,
+            color: item.blend ? backgroundColor : null,
+          ),
+        ),
+        curve: curve,
       );
     }
     return Container(
-      color: Colors.transparent,
       padding: EdgeInsets.only(bottom: 2),
       child: Column(
         mainAxisAlignment: MainAxisAlignment.end,